style(domains): co-locate race condition test in domains_test.py
Move test_delete_parent_domain_immediately_after_child_deletion from its
own file into the existing domains_test.py, keeping all domain smoke
tests together consistent with the rest of the test suite.
